# swagger-documentation.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"Registration Successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Auth"]}},"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/v1/auth":{"post":{"operationId":"AuthController_confirmEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/api/v1/users":{"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/v1/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]},"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["Users"]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/api/v1/reviews":{"get":{"operationId":"ReviewsController_getAllReviews","parameters":[],"responses":{"200":{"description":""}},"tags":["Reviews"]},"post":{"operationId":"ReviewsController_createReview","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":""}},"tags":["Reviews"]}}},"info":{"title":"NestJs Hackathon Kit","description":"NestJs API Starter Hackthon Kit","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"http://localhost:8989"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"password":{"type":"string"},"passwordConfirm":{"type":"string"}},"required":["name","email","phoneNumber","password","passwordConfirm"]},"User":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"},"phoneNumber":{"type":"string"},"password":{"type":"string"},"isEmailVerified":{"type":"boolean"},"role":{"type":"string"}},"required":["id","name","email","phoneNumber","password","isEmailVerified","role"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","description":"Email address of the user","example":"reachme@amitavroy.com"},"password":{"type":"string","description":"Password in plain text","example":"Password@123"}},"required":["email","password"]},"UpdateUserDto":{"type":"object","properties":{}},"CreateReviewDto":{"type":"object","properties":{}}}}}